backend-python/requirements.txt
flask
flask-cors
python-dotenv
requests
PyPDF2
opencv-python
pytesseract
pillow
numpy
werkzeug
pdf2image